% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sources.R
\name{compare_sources}
\alias{compare_sources}
\title{Compare dietary proportions between multiple sources}
\usage{
compare_sources(simmr_out, source_names = simmr_out$input$source_names,
  plot = TRUE)
}
\arguments{
\item{simmr_out}{An object of class \code{simmr_output} created from
\code{\link{simmr_mcmc}}.}

\item{source_names}{The names of at least two sources. These should match
the names exactly given to \code{\link{simmr_load}}.}

\item{plot}{A logical value specifying whether plots should be produced or
not.}
}
\value{
If there are two sources, a vector containing the differences
between the two dietary proportion proportions for these two sources. If
there are multiple sources, a list containing the following fields:
\item{Ordering }{The different possible orderings of the dietary proportions
across sources} \item{out_all }{The dietary proportions for these sources
specified as columns in a matrix}
}
\description{
This function takes in an object of class \code{simmr_output} and creates
probabilistic comparisons between the supplied sources.
}
\details{
When two sources are specified, the function produces a direct calculation
of the probability that the dietary proportion for one source is bigger than
the other. When more than two sources are given, the function produces a set
of most likely probabilistic orderings for each combination of sources. The
function produces boxplots by default and also allows for the storage of the
output for further analysis if required.
}
\examples{
# Data set 1: 10 obs on 2 isos, 4 sources, with tefs and concdep
# See simmr_mcmc and vignettes for full example run
data(simmr_data_1)

# Load into simmr (includes iso-space plot)
simmr_1 = simmr_data_1 \%>\% 
  simmr_load \%>\% 
  simmr_mcmc
 
# Compare two sources
simmr_1 \%>\% compare_sources(source_names = c('Source B', 'Source D'))

# Compare multiple sources
simmr_1 \%>\% compare_sources

}
\seealso{
See \code{\link{simmr_mcmc}} for complete examples.
}
\author{
Andrew Parnell <andrew.parnell@mu.ie>
}
