% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{generate}
\alias{generate}
\title{A Data-Generating Function}
\usage{
generate(params)
}
\arguments{
\item{params}{List of parameters used to generate simulated data}
}
\description{
This function generates simulated data given a list of parameters
}
\examples{
generate()
}
\keyword{data}
