% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{parthstd}
\alias{parthstd}
\title{Helper function}
\usage{
parthstd(bx, by, bxse, byse, maf, nx, ny, MR_standardization_type)
}
\arguments{
\item{bx}{}
}
\description{
Helper function
}
\examples{
parthstd()
}
