% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simmr_input_ts.R
\name{plot.simmr_input_ts}
\alias{plot.simmr_input_ts}
\title{Plot the \code{simmr_input} data created from \code{simmr_load}}
\usage{
\method{plot}{simmr_input_ts}(x, tracers = c(1, 2),
  title = "Tracers plot", xlab = x$tracer_names[tracers[1]],
  ylab = x$tracer_names[tracers[2]], sigmas = 1,
  mix_name = "Mixtures", ggargs = NULL, colour = TRUE, ...)
}
\arguments{
\item{x}{An object created via the function \code{\link{simmr_load}}}

\item{tracers}{The choice of tracers to plot. If there are more than two
tracers, it is recommended to plot every pair of tracers to determine
whether the mixtures lie in the mixing polygon defined by the sources}

\item{title}{A title for the graph}

\item{xlab}{The x-axis label. By default this is assumed to be delta-13C but
can be made richer if required. See examples below.}

\item{ylab}{The y-axis label. By default this is assumed to be delta-15N in
per mil but can be changed as with the x-axis label}

\item{sigmas}{The number of standard deviations to plot on the source
values. Defaults to 1.}

\item{mix_name}{A optional string containing the name of the mixture
objects, e.g. Geese.}

\item{ggargs}{Extra arguments to be included in the ggplot (e.g. axis limits)}

\item{colour}{If TRUE (default) creates a plot. If not, puts the plot in
black and white}

\item{...}{Not used}
}
\description{
This function creates iso-space (AKA tracer-space or delta-space) plots.
They are vital in determining whether the data are suitable for running in a
SIMM.
}
\details{
It is desirable to have the vast majority of the mixture observations to be
inside the convex hull defined by the food sources. When there are more than
two tracers (as in one of the examples below) it is recommended to plot all
the different pairs of the food sources. See the vignette for further
details of richer plots.
}
\examples{

# Data set 1: 10 obs on 2 isos, 4 sources, with tefs and concdep
# See simmr_mcmc and vignettes for full example run
data(simmr_data_1)

# Load into simmr - automatically produces plot
simmr_1 = simmr_data_1 \%>\% 
  simmr_load \%>\%
  simmr_mcmc
  
# Plot it (again) from the output
simmr_1 \%>\% plot(type = 'isospace')
 
# Or plot it from the input
simmr_1$input \%>\% plot

### A more complicated example with 30 obs, 3 tracers and 4 sources
data(simmr_data_2)

# Load in and plot by default the first two isotopes
simmr_2 = simmr_data_2 \%>\%
            simmr_load()

# To get at other isotopes call plot 3 times
simmr_2 \%>\% plot(type = 'isospace', tracers = c(1, 2))
# Now plot d15N vs d34S
simmr_2 \%>\% plot(type = 'isospace', tracers = c(2, 3))
# and finally d13C vs d34S
simmr_2 \%>\% plot(type = 'isospace', tracers = c(1, 3))
# See vignette('simmr') for fancier x-axis labels

}
\seealso{
See \code{\link{plot.simmr_output}} for plotting the output of a
simmr run. See \code{\link{simmr_mcmc}} for running a simmr object once the
iso-space is deemed acceptable.
}
\author{
Andrew Parnell <andrew.parnell@mu.ie>
}
