% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simmr_output.R
\name{plot.simmr_output}
\alias{plot.simmr_output}
\title{Plot different features of an object created from \code{\link{simmr_mcmc}}.}
\usage{
\method{plot}{simmr_output}(x, type = c("isospace", "histogram",
  "density", "matrix", "boxplot"), binwidth = 0.05, alpha = 0.5,
  title = "simmr output plot", ggargs = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{simmr_output} created via
\code{\link{simmr_mcmc}}}

\item{type}{The type of plot required. Can be one or more of 'histogram',
'density', 'matrix', or 'boxplot'}

\item{binwidth}{The width of the bins for the histogram. Defaults to 0.05}

\item{alpha}{The degree of transparency of the plots. Not relevant for
matrix plots}

\item{title}{The title of the plot.}

\item{ggargs}{Extra arguments to be included in the ggplot (e.g. axis limits)}

\item{...}{Currently not used}
}
\description{
This function allows for 4 different types of plots of the simmr output
created from \code{\link{simmr_mcmc}}. The types are: histogram, kernel
density plot, matrix plot (most useful) and boxplot. There are some minor
customisation options.
}
\details{
The matrix plot should form a necessary part of any SIMM analysis since it
allows the user to judge which sources are identifiable by the model.
Further detail about these plots is provided in the vignette.
}
\examples{

# Data set 1: 10 obs on 2 isos, 4 sources, with tefs and concdep
# See simmr_mcmc and vignettes for full example run
data(simmr_data_1)

# Load into simmr and run
simmr_1 = simmr_data_1 \%>\% 
  simmr_load \%>\% 
  simmr_mcmc

# Plot output
simmr_1 \%>\% plot(type='boxplot')
simmr_1 \%>\% plot(type='histogram')
simmr_1 \%>\% plot(type='density')
simmr_1 \%>\% plot(type='matrix') # Often the most useful
}
\seealso{
See \code{\link{simmr_mcmc}} for creating objects suitable for this
function, and many more examples. See also \code{\link{simmr_load}} for
creating simmr objects, \code{\link{plot.simmr_input}} for creating isospace
plots, \code{\link{summary.simmr_output}} for summarising output.
}
\author{
Andrew Parnell <andrew.parnell@mu.ie>
}
