% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simmr_input.R
\name{print.simmr_input}
\alias{print.simmr_input}
\title{Print simmr input object}
\usage{
\method{print}{simmr_input}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simmr_input}}

\item{...}{Other arguments (not supported)}
}
\value{
A neat presentation of your simmr object.
}
\description{
Print simmr input object
}
