% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simmr_input_ts.R
\name{print.simmr_input_ts}
\alias{print.simmr_input_ts}
\title{Print simmr input time series object}
\usage{
\method{print}{simmr_input_ts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simmr_input_ts}}

\item{...}{Other arguments (not supported)}
}
\value{
A neat presentation of your simmr object.
}
\description{
Print simmr input time series object
}
