% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simmr_output_re.R
\name{print.simmr_output_re}
\alias{print.simmr_output_re}
\title{Print a simmr output object}
\usage{
\method{print}{simmr_output_re}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simmr_output}}

\item{...}{Other arguments (not supported)}
}
\value{
Returns a neat summary of the object
}
\description{
Print a simmr output object
}
\seealso{
\code{\link{simmr_mcmc}} for creating \code{simmr_output} objects
}
