% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr_elicit_basic.R
\name{simmr_elicit_basic}
\alias{simmr_elicit_basic}
\title{Function to allow informative prior distribution to be included in simmr}
\usage{
simmr_elicit_basic(n_sources, proportion_means = rep(1/n_sources,
  n_sources), proportion_sds = rep(0.1, n_sources), n_sims = 1000)
}
\arguments{
\item{n_sources}{The number of sources required}

\item{proportion_means}{The desired prior proportion means. These should sum
to 1. Should be a vector of length \code{n_sources}}

\item{proportion_sds}{The desired prior proportions standard deviations.
These have no restricted sum but should be reasonable estimates for a
proportion.}

\item{n_sims}{The number of simulations for which to run the optimsiation
routine.}
}
\value{
A list object with two components \item{mean }{The best estimates of
the mean to use in \code{control.prior} in \code{\link{simmr_mcmc}}}
\item{sd }{The best estimates of the standard deviations to use in
\code{control.prior} in \code{\link{simmr_mcmc}}}
}
\description{
The main \code{\link{simmr_mcmc}} function allows for a prior distribution
to be set for the dietary proportions. The prior distribution is specified
by transforming the dietary proportions using the centralised log ratio
(CLR). The \code{\link{simmr_elicit}} and \code{\link{simmr_elicit_basic}}
functions allows the user to specify
prior means and standard deviations for each of the dietary proportions, and
then finds CLR-transformed values suitable for input into
\code{\link{simmr_mcmc}}. Th version \code{\link{simmr_elicit_basic}} differs
from \code{\link{simmr_elicit}} because it does not require a loaded 
\code{simmr} object so can be used for e.g. simulation purposes
}
\details{
The function takes the desired proportion means and standard deviations,
and fits an optimised least squares to the means and standard deviations in
turn to produced CLR-transformed estimates for use in
\code{\link{simmr_mcmc}}. Using prior information in SIMMs is highly
desirable given the restricted nature of the inference. The prior
information might come from previous studies, other experiments, or other
observations of e.g. animal behaviour.

Due to the nature of the restricted space over which the dietary proportions
can span, and the fact that this function uses numerical optimisation, the
procedure will not match the target dietary proportion meanss and standard
deviations exactly. If this problem is severe, try increasing the
\code{n_sims} value.
}
\examples{

\dontrun{
# Suppose we had prior information that:
# proportion means = 0.5,0.2,0.2,0.1
# proportion sds = 0.08,0.02,0.01,0.02
prior=simmr_elicit_basic(4,c(0.5,0.2,0.2,0.1),c(0.08,0.02,0.01,0.02))
}
}
\seealso{
See \code{\link{simmr_elicit}} to use with an existing \code{simmr}
object.
}
\author{
Andrew Parnell <andrew.parnell@ucd.ie>
}
