% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{square_data}
\alias{square_data}
\title{An artificial data set used to indicate effect of priors}
\format{
A list with the following elements
\describe{
  \item{mixtures}{A two column matrix containing delta 13C and delta 15N values respectively}
  \item{source_names}{A character vector of the food source names}
  \item{tracer_names}{A character vector of the tracer names (d13C, d15N)}
  \item{source_means}{A matrix of source mean values for the tracers in the same order as \code{mixtures} above}
  \item{source_sds}{A matrix of source sd values for the tracers in the same order as \code{mixtures} above}
  \item{correction_means}{A matrix of TEFs mean values for the tracers in the same order as \code{mixtures} above}
  \item{correction_sds}{A matrix of TEFs sd values for the tracers in the same order as \code{mixtures} above}
  \item{concentration_means}{A matrix of concentration dependence mean values for the tracers in the same order as \code{mixtures} above}
  ...

  @seealso \code{\link{simmr_mcmc}} for an example where it is used


}
}
\source{
<doi:10.3354/meps10535>
}
\usage{
square_data
}
\description{
A fake box data set identified by Fry (2014) as a failing of SIMMs
See the link for more interpretation of these data and the output
}
\keyword{datasets}
