% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simmr_output.R
\name{summary.simmr_output}
\alias{summary.simmr_output}
\title{Summarises the output created with \code{\link{simmr_mcmc}} or 
\code{\link{simmr_ffvb}}}
\usage{
\method{summary}{simmr_output}(
  object,
  type = c("diagnostics", "quantiles", "statistics", "correlations"),
  group = 1,
  ...
)
}
\arguments{
\item{object}{An object of class \code{simmr_output} produced by the
function \code{\link{simmr_mcmc}} or \code{\link{simmr_ffvb}}}

\item{type}{The type of output required. At least none of 'diagnostics',
'quantiles', 'statistics', or 'correlations'.}

\item{group}{Which group or groups the output is required for.}

\item{...}{Not used}
}
\value{
A list containing the following components: \item{gelman }{The
convergence diagnostics} \item{quantiles }{The quantiles of each parameter
from the posterior distribution} \item{statistics }{The means and standard
deviations of each parameter} \item{correlations }{The posterior
correlations between the parameters} Note that this object is reported
silently so will be discarded unless the function is called with an object
as in the example below.
}
\description{
Produces textual summaries and convergence diagnostics for an object created
with \code{\link{simmr_mcmc}} or \code{\link{simmr_ffvb}}. The different
options are: 'diagnostics' which produces Brooks-Gelman-Rubin diagnostics
to assess MCMC convergence, 'quantiles' which produces credible intervals
for the parameters, 'statistics' which produces means and standard
deviations, and 'correlations' which produces correlations between the
parameters.
}
\details{
The quantile output allows easy calculation of 95 per cent credible
intervals of the posterior dietary proportions. The correlations, along with
the matrix plot in \code{\link{plot.simmr_output}} allow the user to judge
which sources are non-identifiable. The Gelman diagnostic values should be
close to 1 to ensure satisfactory convergence.

When multiple groups are included, the output automatically includes the
results for all groups.
}
\examples{
\donttest{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)
simmr_1 <- with(
  geese_data_day1,
  simmr_load(
    mixtures = mixtures,
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)

# Plot
plot(simmr_1)


# MCMC run
simmr_1_out <- simmr_mcmc(simmr_1)

# Summarise
summary(simmr_1_out) # This outputs all the summaries
summary(simmr_1_out, type = "diagnostics") # Just the diagnostics
# Store the output in an
ans <- summary(simmr_1_out,
  type = c("quantiles", "statistics")
)
}
}
\seealso{
See \code{\link{simmr_mcmc}} and \code{\link{simmr_ffvb}}for
creating objects suitable for this function, and many more examples.
See also \code{\link{simmr_load}} for creating simmr objects,
\code{\link{plot.simmr_input}} for creating isospace plots,
\code{\link{plot.simmr_output}} for plotting output.
}
\author{
Andrew Parnell <andrew.parnell@mu.ie>, Emma Govan
}
