% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simmr_input_ts.R
\name{plot.simmr_input_ts}
\alias{plot.simmr_input_ts}
\title{Plot a simmr time series}
\usage{
\method{plot}{simmr_input_ts}(x, tracers = c(1, 2),
  title = "Tracers plot", xlab = x$tracer_names[tracers[1]],
  ylab = x$tracer_names[tracers[2]], sigmas = 1,
  mix_name = "Mixtures", time_units = "Time", ggargs = NULL,
  colour = TRUE, ...)
}
\arguments{
\item{x}{An object created via the function \code{\link{simmr_load}}}

\item{tracers}{The choice of tracers to plot. If there are more than two
tracers, it is recommended to plot every pair of tracers to determine
whether the mixtures lie in the mixing polygon defined by the sources}

\item{title}{A title for the graph}

\item{xlab}{The x-axis label. By default this is assumed to be delta-13C but
can be made richer if required. See examples below.}

\item{ylab}{The y-axis label. By default this is assumed to be delta-15N in
per mil but can be changed as with the x-axis label}

\item{sigmas}{The number of standard deviations to plot on the source
values. Defaults to 1.}

\item{mix_name}{A optional string containing the name of the mixture
objects, e.g. Geese.}

\item{ggargs}{Extra arguments to be included in the ggplot (e.g. axis limits)}

\item{colour}{If TRUE (default) creates a plot. If not, puts the plot in
black and white}

\item{...}{Not used}
}
\description{
This function creates iso-space (AKA tracer-space or delta-space) plots for time series models.
They are vital in determining whether the data are suitable for running in a
SIMM.
}
\examples{

}
\seealso{
See \code{\link{plot.simmr_input}} for plotting standard simmr objects. See \code{\link{simmr_mcmc}} for running a time series simmr object once the iso-space is deemed acceptable.
}
\author{
Andrew Parnell <andrew.parnell@mu.ie>
}
