% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_viz.simmr_output.R
\name{prior_viz}
\alias{prior_viz}
\title{Plot the prior distribution for a simmr run}
\usage{
prior_viz(simmr_out, plot = TRUE, include_posterior = TRUE,
  n_sims = 10000)
}
\arguments{
\item{simmr_out}{A run of the simmr model from \code{\link{simmr_mcmc}}}

\item{plot}{Whether to create a density plot of the prior or not. The simulated prior values are returned as part of the object}

\item{include_posterior}{Whether to include the posterior distribution on top of the priors. Defaults to TRUE}

\item{n_sims}{The number of simulations from the prior distribution}
}
\description{
This function takes the output from \code{\link{simmr_mcmc}} and plots the prior distribution to enable visual inspection. This can be used by itself or as part of \code{\link{posterior_predictive}} to visually evaluate the influence of the prior on the posterior distribution.
}
\examples{
# Data set 1: 10 obs on 2 isos, 4 sources, with tefs and concdep
# See simmr_mcmc and vignettes for full example run
data(simmr_data_1)

# Load into simmr and run
simmr_1 = simmr_data_1 \%>\% 
  simmr_load \%>\% 
  simmr_mcmc

# Prior and posterior distributions
prior = simmr_1 \%>\% prior_viz
head(prior)
summary(prior)
}
