% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{simmr}
\alias{simmr}
\title{Similarity-based multimodal regression}
\usage{
simmr(
  D,
  X,
  variables,
  tests = c("Dempster", "PC"),
  D.scale = c("max", "trace", "dvar", "none"),
  pc.pv = NULL,
  pc.n = NULL,
  n.perm = 999,
  eigen.tol = 1e-10,
  debug = FALSE
)
}
\arguments{
\item{D}{List of distance matrices of class \link[base]{matrix} or
\link[stats]{dist}}

\item{X}{Design matrix, either output of \link[stats]{model.matrix} or valid
input to the function.}

\item{variables}{Names of explanatory variables to regress on, must match
with column names of \code{X}}

\item{tests}{Chosen test statistics from "Dempster", "PC". For "PC", defaults
to using n-p-1 PCs, where \code{X} is \verb{n x p}. If \code{pc.pv} and \code{pc.n} are both
specified, \code{pc.pv} is prioritized.}

\item{D.scale}{Method for normalization of distance matrices}

\item{pc.pv}{Optional, calculate additional PC-based test statistics
for specified proportions of variation explained}

\item{pc.n}{Optional, calculate additional PC-based test statistics
for specific numbers of PCs. If 'all', tries every possible number of PCs.}

\item{n.perm}{Number of permutations to perform}

\item{eigen.tol}{Discard eigenvalues below this tolerance for both PCA and
computing test statistics}

\item{debug}{Include internal objects in output (from unpermuted analysis)}
}
\value{
A list with the following components:
\item{\code{stat}}{Test statistics computed on original data.}
\item{\code{perms}}{Test statistics after permutations.}
\item{\code{perm.p}}{Permutation p-values for each test.}
}
\description{
\code{simmr} fits a multimodal regression model using only distance matrices
computed on the original data. Permutation testing is used to obtain
\emph{p}-values.
}
\examples{
D <- list(dist(rnorm(10)), dist(rnorm(10)), dist(rnorm(10)))
X <- list("var" = runif(10))
simmr(D, X, "var")
}
