% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr_load.R
\name{simmr_load}
\alias{simmr_load}
\title{Load data into simmr and check for errors}
\usage{
simmr_load(data, fixed_covars = NULL, random_covars = NULL,
  time_var = NULL, print = TRUE, iso_plot = TRUE)
}
\arguments{
\item{data}{A named list with the elements \code{mixtures}, \code{source_names}, \code{source_means}, and \code{source_sds}. Optionally \code{correction_means}, \code{correction_sds}, and \code{concentration_means} can be provided. \code{simmr_load} will provide a note if these are not used.}

\item{fixed_covars}{A matrix of covariates to be used as fixed effects. Should have same number of rows as \code{data$mixtures}.}

\item{random_covars}{A matrix of covariates to be used as random effects. Should have same number of rows as \code{data$mixtures}.}

\item{time_var}{A vector of time values (may be discrete or continuous) to use in a time series analysis. Should the same length as the number of rows as \code{data$mixtures}.}
}
\value{
A named list which is suitable for input into \link{\code{plot.simmr_input}}, \link{\code{simmr_mcmc}} and other fitting functions.
}
\description{
Load data into simmr and check for errors
}
\examples{
#' # A simple example with 10 observations, 2 tracers and 4 sources

data = list(
mixtures = matrix(c(-10.13, -10.72, -11.39, -11.18, -10.81, -10.7, -10.54,
-10.48, -9.93, -9.37, 11.59, 11.01, 10.59, 10.97, 11.52, 11.89,
11.73, 10.89, 11.05, 12.3), ncol=2, nrow=10),
source_names=c('Source A','Source B','Source C','Source D'),
tracer_names = c('d13C','d15N'),
source_means = matrix(c(-14, -15.1, -11.03, -14.44, 3.06, 7.05, 13.72, 5.96), ncol=2, nrow=4),
source_sds = matrix(c(0.48, 0.38, 0.48, 0.43, 0.46, 0.39, 0.42, 0.48), ncol=2, nrow=4),
correction_means = matrix(c(2.63, 1.59, 3.41, 3.04, 3.28, 2.34, 2.14, 2.36), ncol=2, nrow=4),
correction_sds = matrix(c(0.41, 0.44, 0.34, 0.46, 0.46, 0.48, 0.46, 0.66), ncol=2, nrow=4),
concentration_means = matrix(c(0.02, 0.1, 0.12, 0.04, 0.02, 0.1, 0.09, 0.05), ncol=2, nrow=4))

# Load in with simmr_load
simmr_1 = data \%>\% simmr_load

print(simmr_1)

}
